YUK-18681 Multidrop DNP3 support for TCP Port Genericized bulk insertion to allow boost::iterator_range Removed bool overload Added explicit conversions for two existing implicit conversion cases (CtiDate to CtiTime, enum to int)

+++ yukon-server/LoadManagement/lmcontrolareastore.cpp
@@ -840,7 +840,7 @@ void CtiLMControlAreaStore::reset()
 
                 Cti::Database::DatabaseReader rdr(connection, sql);
 
-                rdr << CtiTablePaoExclusion::ExFunctionLMSubordination;
+                rdr << static_cast<int>(CtiTablePaoExclusion::ExFunctionLMSubordination);
 
                 rdr.execute();
